Each MFAIR object has a number of slots which store information. Key slots to access are listed below.

Each MFAIR object has a number of slots which store information. Key slots to access are listed below.

Value

MFAIR class.

Slots

Y

A matrix or Matrix::dgCMatrix. The main data matrix of N samples and M features.

X

A data.frame. The auxiliary information data frame of N samples and C covariates.

Y_sparse

Logical. Whether the main data matrix Y is stored in the sparse mode.

Y_center

Logical. Whether the main data matrix Y is centered.

Y_mean

Numeric. Mean of the main data matrix Y if centered. Zero if not.

Y_missing

Logical. Whether the main data matrix Y is partially observed.

n_obs

integer. Total number of observed entries in Y.

N

An integer. Number of rows (samples) of Y, also the number of rows (samples) of X.

M

An integer. Number of columns (features) of Y.

C

An integer. Number of columns (auxiliary covariates) of X.

K_max

An integer. Note that increasing K_max does not ensure that the actual K also increases, since K_max is just an upper bound and the model will automatically infer K below K_max. If you want inference with larger rank K, please make sure that K_max is large enough and the tol_snr parameter in the fitting function fitGreedy() is small enough, simultaneously.

K

An integer. The inferred rank of Y.

Z

An N * K matrix. Estimated loading matrix, corresponding to the inferred posterior mean of Z in the MFAI model.

a_sq

A matrix containing posterior variance of Z with k-th column corresponding to the k-th loading. For fully observed Y, all N elements of one specific loading share the same posterior variance, then a_sq is a 1 * K matrix. For Y with missing data, elements of one specific loading have different posterior variances, then a_sq is an N * K matrix.

W

An M * K matrix. Estimated factor matrix, corresponding to the inferred posterior mean of W in the MFAI model.

b_sq

A matrix containing posterior variance of W with k-th column corresponding to the k-th factor. For fully observed Y, all M elements of one specific factor share the same posterior variance, then b_sq is a 1 * K matrix. For Y with missing data, elements of one specific factor have different posterior variances, then b_sq is an M * K matrix.

tau

Numeric. A vector of length K, containing the precision parameter for each pair of loading/factor.

beta

Numeric. A vector of length K, containing the precision parameter for each loading Z_k.

FX

An N * K matrix representing the prior mean of Z, corresponding to F(X) in the MFAI model.

tree_0

An 1 * K matrix containing tree_0 with k-th column corresponding to the k-th factor. Tree_0 is defined as the mean of mu vector in each factor.

tree_lists

A list of length K, containing K fitted functions and each function is represented as a list of trees, i.e., the k-th list corresponds to function F_k(.) in the MFAI model.

initialization

A list. Initialization of the fitted model.

learning_rate

Numeric. The learning rate in the gradient boosting part.

tree_parameters

A list of options that control details of the rpart algorithm.

project

Character. Name of the project (for record keeping).