MFAIRSingleFactor object contains the key information about the fitted single factor MFAI model.

MFAIRSingleFactor object contains the key information about the fitted single factor MFAI model.

Value

MFAIRSingleFactor class.

Slots

Y_missing

Logical. Whether the main data matrix Y is partially observed.

n_obs

Integer. Total number of observed entries in Y.

mu

An vector of length N representing the inferred loading, corresponding to the posterior mean of z in the single factor MFAI model.

a_sq

Numeric. The posterior variance of the loading z. For fully observed Y, all N elements of the loading share the same posterior variance, then a_sq is a single number. For Y with missing data, the elements have different posterior variances, then a_sq is a vector of length N.

nu

An vector of length M representing the inferred factor, corresponding to the posterior mean of w in the single factor MFAI model.

b_sq

Numeric. The posterior variance of the factor w. For fully observed Y, all M elements of the factor share the same posterior variance, then b_sq is a single number. For Y with missing data, the elements have different posterior variances, then b_sq is a vector of length M.

tau

Numeric. Precision parameter this pair of loading/factor.

beta

Numeric. Precision parameter for this loading z.

FX

An vector of length N representing the prior mean of z, corresponding to F(X) in the single factor MFAI model.

tree_0

Numeric. Tree_0 is defined as the mean of mu vector.

tree_list

A list containing multiple decision trees, corresponding to function F(.) in the single factor MFAI model.

project

Character. Name of the project (for record keeping).