Fit the MFAI model using backfitting algorithm.

Fit the MFAI model using backfitting algorithm.

Usage

fitBack(
  object,
  learning_rate = 0.1,
  minsplit = 10,
  minbucket = round(minsplit/3),
  maxdepth = 2,
  other_tree_para = list(),
  iter_max_bf = 5000,
  tol_bf = 0.01,
  verbose_bf_inner = TRUE,
  verbose_bf_outer = TRUE,
  sf_para = list()
)

Arguments

object

MFAIR object.

learning_rate

Numeric. Parameter for the gradient boosting part.

minsplit

Numeric. Parameter for the gradient boosting part.

minbucket

Numeric. Parameter for the gradient boosting part.

maxdepth

Numeric. Parameter for the gradient boosting part.

other_tree_para

A list containing other parameters for the gradient boosting part. See rpart::rpart.control() for details.

iter_max_bf

Integer. Maximum iterations allowed.

tol_bf

Numeric. The convergence criterion.

verbose_bf_inner

Logical. Whether to display the detailed information during the inner loop.

verbose_bf_outer

Logical. Whether to display the detailed information during the outer loop.

sf_para

A list containing parameters for fitting the single factor MFAI model. See fitSFFully(), fitSFMissing(), or fitSFSparse() for details.

Value

An MFAIR object containing the information about the fitted MFAI model using backfitting algorithm.