Fit the MFAI model using greedy algorithm.

Fit the MFAI model using greedy algorithm.

Usage

fitGreedy(
  object,
  K_max = NULL,
  learning_rate = 0.1,
  minsplit = 10,
  minbucket = round(minsplit/3),
  maxdepth = 2,
  other_tree_para = list(),
  null_check = TRUE,
  tol_snr = 0.002,
  verbose_greedy = TRUE,
  save_init = FALSE,
  sf_para = list()
)

Arguments

object

MFAIR object.

K_max

Integer. The maximum rank allowed in the MFAI model.

learning_rate

Numeric. Parameter for the gradient boosting part.

minsplit

Integer. Parameter for the gradient boosting part.

minbucket

Integer. Parameter for the gradient boosting part.

maxdepth

Integer. Parameter for the gradient boosting part.

other_tree_para

A list containing other parameters for the gradient boosting part. See rpart::rpart.control() for details.

null_check

Logical. If TRUE, then mfair will check whether the currently inferred factor is close to zero compared to the estimated noise strength. If the check is performed and successful, the factor will be deleted in the returned fit, and the greedy algorithm will be terminated.

tol_snr

Numeric. The convergence criterion which determine the inferred rank of data.

verbose_greedy

Logical. Whether to display the detailed information when fitting the model.

save_init

Logical. Whether to save the initialization of the model.

sf_para

A list containing parameters for fitting the single factor MFAI model. See fitSFFully(), fitSFMissing(), or fitSFSparse() for details.

Value

An MFAIR object containing the information about the fitted MFAI model using greedy algorithm.