MovieLens 100K data.

A list containing the movie rating, user information, and movie genres of the MovieLens 100K data.

Usage

ml100k

Format

A list containing a movie rating matrix, a user information data frame, and a movie genres data frame.

rating

Movie rating matrix of 943 users and 1,682 movies (0–5 star rating).

user

A data frame containing the information of 943 users, the three columns correspond to age, gender, and occupation respectively.

genre

A binary data frame containing the genre information of 1,682 movies, each column corresponds to one specific genre.

Source

https://grouplens.org/datasets/movielens/100k/