Update the k-th factor of the MFAIR object in the backfitting algorithm.

Update the k-th factor of the MFAIR object in the backfitting algorithm.

Usage

updateMFAIR(object, object_sf, k)

Arguments

object

An MFAIR object containing the initial information about the K-factor MFAI model.

object_sf

A MFAIRSingleFactor containing the information about the newly fitted single factor MFAI model.

k

Integer. Which fator to be updated.

Value

MFAIR object containing the information about the new fitted single factor.